revamp project to run with pz 41.71
* bump java version to 17
* add a contribute file, to describe how to get started when contributing
* bump gradle wrapper to 7.2
* fix java and zomboid version related problems
* change run configuration to use java 17 zulu
* there seems to be a problem with fmod, skipping hook
  OnLoadSoundBanksHook integration if target null
* checking for java 17 in run scripts
* ignoring jdk.internal.reflect. package on class loading
* fix broken test, urls couldn't be parsed like this anymore
* refactor Asm Code to work with java 17
* remove unnecessary dependency from gralde build
* add instructions before first node instead of before
  first label node
* revert an unnecessary change
* fix OnLoadSoundBanksHook
  the java code has changed in the init method, so I've
  changed the pattern of this method
* update version numbers
* update classpath to match new libraries in PZ
* make LuaEvents work again
  they stopped working, because of a check for a callback in PZ code.
  Now I've added a dummy callback to make sure, LuaEventManager
  really dispatches the event to StormEventDispatcher

+++ api/src/main/java/io/pzstorm/storm/event/lua/OnObjectLeftMouseButtonUpEvent.java
@@ -36,14 +36,14 @@ public class OnObjectLeftMouseButtonUpEvent implements LuaEvent {
 	/**
 	 * Position of mouse along x-axis.
 	 */
-	public final Integer x;
+	public final Double x;
 
 	/**
 	 * Position of mouse along y-axis.
 	 */
-	public final Integer y;
+	public final Double y;
 
-	public OnObjectLeftMouseButtonUpEvent(IsoObject object, Integer x, Integer y) {
+	public OnObjectLeftMouseButtonUpEvent(IsoObject object, Double x, Double y) {
 
 		this.object = object;
 		this.x = x;
